backend_core/requirements.txt
Django~=5.2.10
dotenv~=0.9.9
python-dotenv~=1.2.1
psycopg2-binary
selenium==4.40.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
django-cors-headers
gunicorn
redis
drf-yasg>=1.21.7
django-redis>=5.3.0
Pillow
requests==2.31.0